Upload js scripts
fortune teller for if statement logic, email masker for demonstration of simple string manipulation. 
Obviously didn't check for any edge cases as this is just to get a handling on the basics of javascripting. 
Maybe I'll revisit them in the future but I most likely won't as I can probably just rely on libraries to handling anything I might need in a practical real world use case.